requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
Jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
PyMySQL==1.1.1
chromadb==0.5.5
sentence-transformers==3.0.1
itsdangerous==2.2.0
